services/report-python/requirements.txt
pika==1.3.2
python-dotenv==1.0.0
